pyproject.toml
[project]
name = "keiz"
version = "0.1.0"
description = "RAG backend for Slayubs using LangChain, FAISS, and Ollama"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "langchain>=0.1.0",
    "langchain-community>=0.0.20",
    "langchain-openai>=0.0.5",
    "faiss-cpu>=1.7.4",
    "openai>=1.12.0",
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "python-dotenv>=1.0.0",
    "pypdf>=4.0.0",
    "python-docx>=1.1.0",
    "sentence-transformers>=2.3.0",
    "pydantic>=2.6.0",
    "pydantic-settings>=2.1.0",
    "python-multipart>=0.0.9",
    "aiofiles>=23.2.1",
    "langchain-experimental>=0.4.0",
    "beautifulsoup4>=4.14.2",
    "lxml>=6.0.2",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "httpx>=0.26.0",
    "black>=24.1.0",
    "ruff>=0.2.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.black]
line-length = 100
target-version = ["py310"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
